fix(checker): wide match-exhaustiveness no longer panics past 128 bits
Ty::Bits(n)'s n is the declared scrutinee width (up to MAX_WIDTH), not
a bound on any individual pattern's natural width -- a bits[200] match
can hold a Bits::Wide literal pattern in `seen`, which the "smallest
missing value" scan's Small-only unreachable!() didn't account for.
Skip that precise gap-finding past n < 128 (the only range where every
seen value is guaranteed Small) and fall back to a generic E0601
message instead of pinpointing a value.

Also clears the clippy debt left by the BUG-13 layer 2 squash
(result_large_err on grown_op/shift_amount/eval_src, two redundant
u128-to-u128 casts in emit_verilog/expr.rs) and two pre-existing
prettier formatting drifts in docs/, so the full fmt/clippy/prettier/
markdownlint gate is actually clean.

Found in code review of the BUG-13 layer 2 branch.

+++ crates/mimz-core/src/emit_verilog/expr.rs
@@ -1185,14 +1185,14 @@ impl Emitter<'_> {
                         base: Box::new(scrutinee.clone()),
                         hi: Box::new(Expr {
                             kind: ExprKind::Int {
-                                value: (hi as u128).into(),
+                                value: hi.into(),
                                 raw: hi.to_string(),
                             },
                             span: sp,
                         }),
                         lo: Box::new(Expr {
                             kind: ExprKind::Int {
-                                value: (lo as u128).into(),
+                                value: lo.into(),
                                 raw: lo.to_string(),
                             },
                             span: sp,
